separate options to dedicated file
