+++ load/doc.go
@@ -1,2 +1,3 @@
-// Package load loads OpenAPI specs from different sources like URL, Path, stdin
+// Package load loads OpenAPI specs from different sources like URLs, paths, globs and stdin
+// Optionally, specs can be preprocessed after loading
 package load
