+++ load/spec_info.go
@@ -5,8 +5,6 @@ import (
 	"net/url"
 
 	"github.com/getkin/kin-openapi/openapi3"
-	"github.com/tufin/oasdiff/flatten/allof"
-	"github.com/tufin/oasdiff/flatten/commonparams"
 	"github.com/yargevad/filepathx"
 )
 
@@ -38,46 +36,6 @@ func getVersion(spec *openapi3.T) string {
 	}
 
 	return spec.Info.Version
-}
-
-type Option func(Loader, []*SpecInfo) ([]*SpecInfo, error)
-
-// WithIdentity returns the original SpecInfos
-func WithIdentity() Option {
-	return func(loader Loader, specInfos []*SpecInfo) ([]*SpecInfo, error) {
-		return specInfos, nil
-	}
-}
-
-func GetOption(option Option, enable bool) Option {
-	if !enable {
-		return WithIdentity()
-	}
-	return option
-}
-
-// WithFlattenAllOf returns SpecInfos with flattened allOf
-func WithFlattenAllOf() Option {
-	return func(loader Loader, specInfos []*SpecInfo) ([]*SpecInfo, error) {
-		var err error
-		for _, specInfo := range specInfos {
-			if specInfo.Spec, err = allof.MergeSpec(specInfo.Spec); err != nil {
-				return nil, err
-			}
-		}
-		return specInfos, nil
-	}
-}
-
-// WithFlattenParams returns SpecInfos with Common Parameters combined into operation parameters
-// See here for Common Parameters definition: https://swagger.io/docs/specification/describing-parameters/
-func WithFlattenParams() Option {
-	return func(loader Loader, specInfos []*SpecInfo) ([]*SpecInfo, error) {
-		for _, specInfo := range specInfos {
-			commonparams.Move(specInfo.Spec)
-		}
-		return specInfos, nil
-	}
 }
 
 // NewSpecInfo creates a SpecInfo from a local file path, a URL, or stdin
